[Feat] Add APIs in repository
