+++ app/src/main/java/com/android/hara/data/repository/HaraAloneRepositoryImpl.kt
@@ -1,10 +1,10 @@
 package com.android.hara.data.repository
 
 import com.android.hara.data.datasource.HaraAloneService
-import com.android.hara.data.model.request.*
+import com.android.hara.data.model.request.DecideAloneReqDto
+import com.android.hara.data.model.request.WorryAloneRequestDto
 import com.android.hara.data.model.response.*
 import com.android.hara.domain.repository.HaraAloneRepository
-import timber.log.Timber
 import javax.inject.Inject
 
 // [서버통신 4] HARARepository에서 선언한 interface를 실제로 구현한다
@@ -13,55 +13,31 @@ import javax.inject.Inject
 // view model에 서버통신 성공/실패 시 작업을 정의한다 (HomeViewModel.kt 참고)
 
 class HaraAloneRepositoryImpl @Inject constructor(
-    private val HARAaloneService: HaraAloneService
+    private val HaraAloneService: HaraAloneService
 ) : HaraAloneRepository {
 
-    override suspend fun getAllPost(categoryId: Int): AllPostResDto {
-        return HARAaloneService.showAllPost(categoryId)
-    }
-
-    override suspend fun postVote(request: VoteReqDto): VoteResDto {
-        return HARAaloneService.vote(request)
-    }
-
     override suspend fun getAloneList(isSolved: Int): WorryListResDto {
-        return HARAaloneService.getAloneList(isSolved)
-    }
-
-    override suspend fun getWithList(isSolved: Int): WorryListResDto {
-        return HARAaloneService.getWithList(isSolved)
+        return HaraAloneService.getAloneList(isSolved)
     }
 
     override suspend fun getRandom(): OnesecResDto {
-        return HARAaloneService.getRandom()
+        return HaraAloneService.getRandom()
     }
 
     override suspend fun getLastWorry(): RandomListResDto {
-        return HARAaloneService.getLastWorry()
-    }
-
-    override suspend fun patchDecideWith(decision: DecideWithReqDto): DecisionResDto {
-        return HARAaloneService.patchWithDecision(decision)
+        return HaraAloneService.getLastWorry()
     }
 
     override suspend fun patchDecideAlone(decision: DecideAloneReqDto): DecisionResDto {
-        return HARAaloneService.patchAloneDecision(decision)
-    }
-
-    override suspend fun postWorryWith(request: WorryWithRequestDto): WorryWithResponseDto {
-        Timber.e(request.toString())
-        return HARAaloneService.postWorryWith(request)
+        return HaraAloneService.patchAloneDecision(decision)
     }
 
     override suspend fun postWorryAlone(request: WorryAloneRequestDto): WorryAloneResponseDto {
-        return HARAaloneService.postWorryAlone(request)
-    }
-
-    override suspend fun getDetailWith(worryId: Int): DetailWithResDto {
-        return HARAaloneService.getDetailWith(worryId)
+        return HaraAloneService.postWorryAlone(request)
     }
 
     override suspend fun getDetailAlone(worryId: Int): DetailAloneResDto {
-        return HARAaloneService.getDetailAlone(worryId)
+        return HaraAloneService.getDetailAlone(worryId)
     }
+
 }
